+++ app/service/service_test.go
@@ -181,7 +181,9 @@ func TestReturnsErrorIfErrorIsProducedWhileSendingEmail(t *testing.T) {
 		TransactionsReader: TransactionsReader{
 			LocalCSVReader: mockCSVReader,
 		},
-		EmailSender: mockEmailSender,
+		EmailService: EmailService{
+			EmailSender: mockEmailSender,
+		},
 	}
 
 	mockCSVReader.On("Read", "correct.csv").Return([][]string{{"0", "7/15", "+60.5"}}, nil)
@@ -203,7 +205,9 @@ func TestProcessReturnsNilIfAllCorrect(t *testing.T) {
 		TransactionsReader: TransactionsReader{
 			LocalCSVReader: mockCSVReader,
 		},
-		EmailSender: mockEmailSender,
+		EmailService: EmailService{
+			EmailSender: mockEmailSender,
+		},
 	}
 
 	mockCSVReader.On("Read", "correct.csv").Return([][]string{{"0", "7/15", "+60.5"}}, nil)
@@ -216,49 +220,4 @@ Average credit amount: 60.5
 
 	err := service.Process("correct.csv", "client@mail.com")
 	require.NoError(t, err)
-}
-
-func TestTransactionsPerMonthToString(t *testing.T) {
-	tests := []struct {
-		name string
-		got  []TransactionsPerMonth
-		want string
-	}{
-		{"empty list", []TransactionsPerMonth{}, ""},
-		{"list with 1 transaction in one month", []TransactionsPerMonth{
-			{Month: time.Date(time.Now().Year(), 7, 1, 0, 0, 0, 0, time.UTC), Amount: 1},
-		}, `Number of transactions in July: 1
-`},
-		{"list with multiple transaction in one month", []TransactionsPerMonth{
-			{Month: time.Date(time.Now().Year(), 7, 1, 0, 0, 0, 0, time.UTC), Amount: 2},
-		}, `Number of transactions in July: 2
-`},
-		{"list with multiple months", []TransactionsPerMonth{
-			{Month: time.Date(time.Now().Year(), 7, 1, 0, 0, 0, 0, time.UTC), Amount: 2},
-			{Month: time.Date(time.Now().Year(), 8, 1, 0, 0, 0, 0, time.UTC), Amount: 2},
-		}, `Number of transactions in July: 2
-Number of transactions in August: 2
-`},
-		{"list with multiple months of another year", []TransactionsPerMonth{
-			{Month: time.Date(2023, 7, 1, 0, 0, 0, 0, time.UTC), Amount: 2},
-			{Month: time.Date(2023, 8, 1, 0, 0, 0, 0, time.UTC), Amount: 2},
-		}, `Number of transactions in July 2023: 2
-Number of transactions in August 2023: 2
-`},
-		{"list with multiple years", []TransactionsPerMonth{
-			{Month: time.Date(time.Now().Year(), 7, 1, 0, 0, 0, 0, time.UTC), Amount: 2},
-			{Month: time.Date(2023, 7, 1, 0, 0, 0, 0, time.UTC), Amount: 2},
-		}, `Number of transactions in July: 2
-Number of transactions in July 2023: 2
-`},
-	}
-
-	for _, tt := range tests {
-		// t.Run enables running "subtests", one for each
-		// table entry. These are shown separately
-		// when executing `go test -v`.
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.want, transactionsPerMonthToString(tt.got))
-		})
-	}
 }
